+++ CHANGELOG.md
@@ -2,10 +2,18 @@
 
 ## [Unreleased]
 
+### Fixed
+
+- [TD-4018] Fix path of profile execution
+
 ### Added
 
+- [TD-4108] Download grants CSV
+  - `POST /api/grants/csv`
 - [TD-4076] Support for grant request approval
 - [TD-4113] Jobs change status is included in Audit Events
+- [TD-3953] Cursor in grants search
+- [TD-4114] Update job status after metadata load process
 - [TD-4077] Grant request processing workflow support
 
 ### Changed